requirements.txt

streamlit==1.31.0
matplotlib==3.8.4
tensorflow==2.16.1
h5py==3.11.0
pandas_datareader==0.10.0
streamlit==1.32.0
yfinance==0.2.37
keras==3.1.1
https://stock-price-predition-dbxvhesya5ghb7cenezjnv.streamlit.app/
https://stock-price-predition-dbxvhesya5ghb7cenezjnv.streamlit.app/
